feat: worker separate file and rollup plugin
